+++ tests/Plugins/Emoji/ParserTest.php
@@ -150,16 +150,16 @@ class ParserTest extends Test
 		return [
 			[
 				'😀',
-				'<img alt="😀" class="emoji" draggable="false" src="//cdn.jsdelivr.net/emojione/assets/4.5/png/64/1f600.png">'
+				'<img alt="😀" class="emoji" draggable="false" src="https://twemoji.maxcdn.com/2/svg/1f600.svg">'
 			],
 			[
 				'😀',
-				'<img alt="😀" class="emoji" draggable="false" src="//cdn.jsdelivr.net/emojione/assets/4.5/png/64/1f600.png">',
+				'<img alt="😀" class="emoji" draggable="false" src="https://twemoji.maxcdn.com/2/svg/1f600.svg">',
 				['tagName' => 'EMOTE']
 			],
 			[
 				':copyright::registered:#️⃣',
-				'<img alt=":copyright:" class="emoji" draggable="false" src="//cdn.jsdelivr.net/emojione/assets/4.5/png/64/00a9.png"><img alt=":registered:" class="emoji" draggable="false" src="//cdn.jsdelivr.net/emojione/assets/4.5/png/64/00ae.png"><img alt="#️⃣" class="emoji" draggable="false" src="//cdn.jsdelivr.net/emojione/assets/4.5/png/64/0023-20e3.png">'
+				'<img alt=":copyright:" class="emoji" draggable="false" src="https://twemoji.maxcdn.com/2/svg/a9.svg"><img alt=":registered:" class="emoji" draggable="false" src="https://twemoji.maxcdn.com/2/svg/ae.svg"><img alt="#️⃣" class="emoji" draggable="false" src="https://twemoji.maxcdn.com/2/svg/23-fe0f-20e3.svg">'
 			],
 		];
 	}